+++ src/c-compiler/types/permission.h
@@ -53,6 +53,6 @@ PermAstNode *idPerm;
 void permDclNames();
 PermAstNode *newPermNode(char ptyp, uint16_t flags, AstNode *locker);
 void permPrint(int indent, PermAstNode *node, char* prefix);
-int8_t permGetFlags(AstNode *node);
+uint16_t permGetFlags(AstNode *node);
 
 #endif
